Solution Category and Subcategory

+++ src/main/java/com/icesi/store/finalproyect/repositories/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends CrudRepository<UserApp, Integer> {
+public interface UserRepository extends CrudRepository<UserApp, Long> {
 
 	UserApp findByUsername(String username);
